Add static cast for `client_results.size()` so it's not ambiguous.
Without this, some compilers complain with:
```
In file included from tensorflow_federated/cc/core/impl/executors/composing_executor_test.cc:31:
./tensorflow_federated/cc/core/impl/executors/value_test_utils.h:83:22: error: call to constructor of 'tensorflow::Tensor' is ambiguous
  tensorflow::Tensor tensor(tensor_constructor_args...);
                     ^      ~~~~~~~~~~~~~~~~~~~~~~~
tensorflow_federated/cc/core/impl/executors/composing_executor_test.cc:510:26: note: in instantiation of function template specialization 'tensorflow_federated::testing::TensorV<unsigned long>' requested here
      v0::Value result = TensorV(client_results.size());
                         ^
```

PiperOrigin-RevId: 407829908

+++ tensorflow_federated/cc/core/impl/executors/composing_executor_test.cc
@@ -507,7 +507,7 @@ TEST_F(ComposingExecutorTest, CreateCallFederatedEvalAtClients) {
     const auto& child = mock_children_[i];
     std::vector<v0::Value> child_results;
     for (uint32_t j = 0; j < clients_per_child_[i]; j++) {
-      v0::Value result = TensorV(client_results.size());
+      v0::Value result = TensorV(static_cast<int32_t>(client_results.size()));
       child_results.push_back(result);
       client_results.push_back(result);
     }
